系统完善

+++ glsx-ne-shield-usercenter/src/main/java/cn/com/glsx/neshield/modules/model/view/DepartmentCount.java
@@ -10,6 +10,6 @@ public class DepartmentCount {
 
     private Long departmentId;
 
-    private Long userNumber;
+    private Integer userNumber;
 
 }
