+++ glsx-ne-shield-usercenter/src/main/java/cn/com/glsx/neshield/modules/model/view/DepartmentUserCount.java
@@ -12,6 +12,6 @@ public class DepartmentUserCount {
 
     private Long departmentId;
 
-    private Long userNumber;
+    private Integer userNumber;
 
 }
